#IP-0749 Fixed #IP-1733 Fixed CONFIG   Removed: `.htgroup' and `.htpasswd' files from the skeleton directory; These files are now created only when needed FRONTEND   Added: Setting that allows administrator to protect/unrpotect default mail accounts against both edition and deletion   Cosmetics: Make usage of the mathematical infinity symbol (∞) in place of the `Unlimited' translation string   Enhancement: Make customers able to edit catch-all accounts   Enhancement: Make customers able to edit mail autoresponder message even if the autoresponder is not activated   Enhancement: Make customers able to select more than one catch-all address in catch-all addresses in drop-down list   Enhancement: Make customers able to show/hide default mail accounts   Enhancement: Protect default mail accounts against change and deletion (default)   Enhancement: Show Catch-all accounts in client mail accounts overview interface   Fixed: Missing creation of default `webmaster' mail account for subdomains   Fixed: Wrong default mail accounts accounting (missing hostmaster email, wrong SQL queries...)   Removed: Information about total items/limits assigned - People don't understand their meaning (statistics)   Review: abuse, hostmaster and postmaster default mail accounts are now forwarded to system admin email (default)   Review: Catchall mail accounts are now counted in mail accounts limit   Review: Default mail accounts are not longer counted for the mail accounts limit (default)   Review: Default mail accounts are now hidden in the client mail accounts overview interface (default)   Review: Make usage of short syntax for arrays   Review: Skip the intermediate edit page when mail autoresponder is being enabled and that message is already set   CS fixes +++ INSTALLER   Fixed: Removal of obsolete files must be done before saving the persistent data, else some files won't be deleted SERVERS   Fixed: `.htgroup' and `.htpasswd' files are reseted when the main domain is being changed (Httpd servers)

+++ gui/public/reseller/software_delete.php
@@ -18,66 +18,37 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-// Include core library
 require 'imscp-lib.php';
 
 iMSCP_Events_Aggregator::getInstance()->dispatch(iMSCP_Events::onResellerScriptStart);
-
 check_login('reseller');
 
-resellerHasFeature('aps') or showBadRequestErrorPage();
-
-/** @var $cfg iMSCP_Config_Handler_File */
-$cfg = iMSCP_Registry::get('config');
-
-if (isset($_GET['id']) AND is_numeric($_GET['id'])) {
-	$query="
-		SELECT
-			`software_id`,
-			`software_archive`,
-			`software_depot`
-		FROM
-			`web_software`
-		WHERE
-			`software_id` = ?
-		AND
-			`reseller_id` = ?
-	";
-	$rs = exec_query($query, [$_GET['id'], $_SESSION['user_id']]);
-
-	if ($rs->recordCount() != 1) {
-		set_page_message(tr('Wrong software id.'), 'error');
-		redirectTo('software_upload.php');
-	} else {
-		if ($rs->fields['software_depot'] == "no") {
-			$del_path = $cfg->GUI_APS_DIR."/".$_SESSION['user_id']."/".$rs->fields['software_archive']."-".$rs->fields['software_id'].".tar.gz";
-			@unlink($del_path);
-		}
-
-		$update = "
-			UPDATE
-				`web_software_inst`
-			SET
-				`software_res_del` = 1
-			WHERE
-				`software_id` = ?
-		";
-		$res = exec_query($update, $rs->fields['software_id']);
-
-		$delete="
-			DELETE FROM
-				`web_software`
-			WHERE
-				`software_id` = ?
-			AND
-				`reseller_id` = ?
-		";
-		$res = exec_query($delete, [$_GET['id'], $_SESSION['user_id']]);
-
-		set_page_message(tr('Software successfully scheduled for deletion.'), 'success');
-		redirectTo('software_upload.php');
-	}
-} else {
-	set_page_message(tr('Wrong software id.'), 'error');
-	redirectTo('software_upload.php');
+if (!resellerHasFeature('aps') || !isset($_GET['id'])) {
+    showBadRequestErrorPage();
 }
+
+$softwareId = intval($_GET['id']);
+$stmt = exec_query(
+    '
+        SELECT software_id, software_archive, software_depot
+        FROM web_software
+        WHERE software_id = ?
+        AND reseller_id = ?
+    ',
+    [$softwareId, $_SESSION['user_id']]
+);
+
+if (!$stmt->rowCount()) {
+    showBadRequestErrorPage();
+}
+
+if ($stmt->fields['software_depot'] == 'no') {
+    $del_path = iMSCP_Registry::get('config')['GUI_APS_DIR'] . '/' . $_SESSION['user_id'] . '/'
+        . $rs->fields['software_archive'] . '-' . $rs->fields['software_id'] . '.tar.gz';
+    @unlink($del_path);
+}
+
+exec_query('UPDATE web_software_inst SET software_res_del = 1 WHERE software_id = ?', $rs->fields['software_id']);
+exec_query('DELETE FROM web_software WHERE software_id = ? AND reseller_id = ?', [$softwareId, $_SESSION['user_id']]);
+set_page_message(tr('Software successfully scheduled for deletion.'), 'success');
+redirectTo('software_upload.php');
